+++ src/main/java/com/example/exoevalcabinet/repositories/DeplacementRepository.java
@@ -3,5 +3,5 @@ package com.example.exoevalcabinet.repositories;
 import com.example.exoevalcabinet.models.Deplacement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeplacementRepository extends JpaRepository<Deplacement,Long> {
+public interface DeplacementRepository extends JpaRepository<Deplacement,Integer> {
 }
